restaurant/requirements.txt
asgiref==3.5.0
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.13.0
django-filter==22.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.7.1
platformdirs==2.5.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.15.1
